Ns-TMのPMLに着手
完全ではないが、一応動く
TEモードは全く手つけてない
MiePrintとNTFの軸移動を修正、大幅に結果改善

+++ Field.h
@@ -35,11 +35,11 @@ public:
 	}
 
 	//nm単位の物理量をシミュレーション値に変換
-	int nanoToCell(const int &length){
+	double nanoToCell(const double &length){
 		return length/H_U;
 	}
 
-	int cellToNano(const int &cell){
+	double cellToNano(const double &cell){
 		return cell*H_U;
 	}
